feat: merge acapella page and music page
close #20
